pyproject.toml
[tool.poetry]
name = "social-media"
version = "0.1.0"
description = "Social media"
authors = ["Ansima"]

[tool.poetry.dependencies]
python = "3.10.11"
fastapi = "^0.95.1"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
SQLAlchemy = "^1.4.39"
databases = "^0.7.0"
python-dotenv = "^1.0.0"
asyncpg = "^0.27.0"
alembic = "^1.10.4"
greenlet = "^2.0.2"
PyJWT = "^2.6.0"
bcrypt = "^4.0.1"
isort = "^5.12.0"
black = "^23.3.0"
Faker = "^18.9.0"
gunicorn = "^20.1.0"
aioredis = "^2.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99
target-version = ["py310"]
skip-string-normalization = true
can-omit-invisible-parents = false

[tool.isort]
profile = "black"
line_length = 99
multi_line_output = 3